+++ clean_python/fastapi/fastapi_access_logger.py
@@ -2,7 +2,6 @@
 
 import os
 import time
-from datetime import datetime
 from typing import Awaitable
 from typing import Callable
 from typing import Optional
@@ -12,6 +11,7 @@ from starlette.background import BackgroundTasks
 from starlette.requests import Request
 from starlette.responses import Response
 
+from clean_python import ctx
 from clean_python import Gateway
 from clean_python.fluentbit import FluentbitGateway
 
@@ -42,10 +42,6 @@ class FastAPIAccessLogger:
             log_access, self.gateway, request, response, time_received, request_time
         )
         return response
-
-
-def fmt_timestamp(timestamp: float) -> str:
-    return datetime.utcfromtimestamp(timestamp).isoformat() + "Z"
 
 
 async def log_access(
@@ -81,7 +77,8 @@ async def log_access(
         "status": response.status_code,
         "content_type": response.headers.get("content-type"),
         "content_length": content_length,
-        "time": fmt_timestamp(time_received),
+        "time": time_received,
         "request_time": request_time,
+        "correlation_id": str(ctx.correlation_id) if ctx.correlation_id else None,
     }
     await gateway.add(item)